requirements.txt
robotframework==6.1.1
robotframework-databaselibrary==1.2.4
ibm-db==3.1.4
ibm-db-sa==0.3.8
sqlalchemy==2.0.23
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0